Prison Break problem.
Find the minimal size battery needed to disable a series of switches if you can also use recharge stations (but only once). This one is also basically a graph problem, but quite tough, since the state of the graph changes depending on the paths you take (charging stations can only be used once).